Make build model parameter computation more discoverable
